OTA streaming: fix image signature verification failure
Issue was happening because during this last write operation block size length
was specified which longer than the actual data length. The later erase action will
start after actual write size not after the actual data. So the actual image length
became bigger and it made the avb footer not located at the end of the image.